UPSTREAM: <carry>: Tolerate node ExternalID changes with no cloud provider
Previously, if the a kubelet tried to register itself the API server,
and was rejected due to the external ID chaning, it would delete the
node object and recreate it.  This commit causes it to tolerate
a change in ExternalID when the ExternalID is not being provided by a
cloud provider, assuming the new ExternalID is either the node's
(metadata) name, or one of node's addresses.