Introduce the Unidler Socket
This commit introduces a variation on the proxysocket handler
(which is responsible for proxying socket traffic in the userspace
proxy).

For TCP sockets, the unidler socket acts normally as
the userspace proxy, except that when it receives connections, it sends
out a `NeedPods` event, and holds the connections until endpoints are
available (or the connection has waited for a certain amount of time).
Once endpoints are available, it acts like a normal TCP proxy socket
for those connections (but does not continue to accept other
connections).

For UDP sockets, the unidler socket drops all UDP traffic it receives,
but periodically sends out `NeedPods` events as long as it continues to
receive traffic.

In either case, once endpoints are available, all new connections should
be routed through the "normal" proxy by using the hybrid proxy.