Make HAProxy Router Aware of Idled Services
This commit makes the HAProxy router aware of unidled services, such
when a service is idled, the HAProxy router will act as if a single
endpoint, pointing to the service IP, is present.  This way, routes
to idled services will be able to take advantage of the unidling service
proxy.