Fork Kubernetes Userspace Proxy
This introduces the "pkg/proxy/userspace" package, which contains
a fork of the userspace proxy code from Kubernetes (cleaned up to pass
go vet).

Additionally, the fork was modified to more readily enable custom
proxy sockets to be used with it (this involves making several
interfaces and fields public).

This can be removed once kubernetes/kubernetes#29465 lands.