Changed the userspace proxy to remove conntrack entries
This changes the userspace proxy so that it cleans up its conntrack
settings when a service is removed (as the iptables proxy already
does).  This could theoretically cause problems when a UDP service was
deleted and recreated quickly (with the same IP address).  As long as
packets from the same UDP source IP and port were going to the same
destination IP and port, then the conntrack would apply and the
packets would be sent to the old destination.

This is astronomically unlikely if you did not specify the IP address
to use in the service, and even then, only happens with an
"established" UDP connection.  However, when unidling a service, we
change from the unidling proxy (based on the userspace proxy) to
iptables (or the plain userspace) and the UDP "connection" that awoke
the service is expected to continue sending... so it frequently would
exercise this case.